+++ src/getBasename.ts
@@ -1,3 +1,3 @@
-import { getIcestark } from './constant';
+import { getIcestark } from './common';
 
 export default () => (getIcestark('basename') ? getIcestark('basename') : '/');
